pyproject.toml
[project]
name = "tspl-netdriver"
version = "0.1.0"
description = "REST TSPL printer driver with web UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pillow",
    "pyudev>=0.24.0",
    "pyusb>=1.2.1",
]

[project.scripts]
tspl-driver = "api.app:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["api", "app_logging", "config", "printer", "usb_access"]
py-modules = [
    "main",
]
